put word32 in end record
this postpones conversion to int untill needed.
allowing us to not have to deal with negatives

+++ src/Zip/Codec/Write.hs
@@ -59,8 +59,8 @@ sinkFile existingCentralDir end zipPath filePath options =
 updateEnd :: DataDescriptor -> FileHeader -> End -> End
 updateEnd dd fh end = end {
     endCentralDirectoryOffset = endCentralDirectoryOffset end
-                            + fromIntegral (localFileHeaderLength fh + ddCompressedSize dd)
-  , endCentralDirectorySize = endCentralDirectorySize end + fromIntegral (fileHeaderLength fh)
+                            + (localFileHeaderLength fh + ddCompressedSize dd)
+  , endCentralDirectorySize = endCentralDirectorySize end + (fileHeaderLength fh)
   }
 
 updateFileHeader :: DataDescriptor -> FileHeader -> FileHeader
